변경: Champion 의 number 를 Long 타입으로 변경

+++ src/main/java/com/neukrang/citadel/lol/domain/ChampionInfo.java
@@ -15,6 +15,6 @@ public class ChampionInfo {
     @JsonIgnoreProperties(ignoreUnknown = true)
     public static class Champion {
         private String id;
-        private Integer key;
+        private Long key;
     }
 }
